Rework new visitor to also handle context reference extraction
